menu and other details

+++ app/src/main/java/com/example/weatherapp/models/Clouds.kt
@@ -3,5 +3,5 @@ package com.example.weatherapp.models
 import java.io.Serializable
 
 data class Clouds (
-    val all: Int
+    val all: Float
     ):Serializable
